Moved code to generate reports into business layer
